+++ CitromTech/CitromTech/src/Renderer/RenderAPI/OpenGL/GLDevice.h
@@ -30,7 +30,7 @@ namespace Citrom::RenderAPI
 		VSyncMode GetVSync() override{return VSyncMode::On;}
 
 		void Resize(uint32 width, uint32 height) override{}
-		void ResizeViewport(uint32 width, uint32 height, int32 xPos = 0, int32 yPos = 0) override{}
+		void ResizeViewport(float32 width, float32 height, int32 xPos = 0, int32 yPos = 0) override{}
 
 		// Buffer
 		VertexBuffer CreateVertexBuffer(VertexBufferDesc* descriptor) override{return VertexBuffer();}
